+++ app/app_upgrades.go
@@ -11,7 +11,7 @@ import (
 	upgrade2_0_0 "github.com/archway-network/archway/app/upgrades/2_0_0"
 	upgrade3_0_0 "github.com/archway-network/archway/app/upgrades/3_0_0"
 	upgrade4_0_0 "github.com/archway-network/archway/app/upgrades/4_0_0"
-	upgrade4_0_1 "github.com/archway-network/archway/app/upgrades/4_0_1"
+	upgrade4_0_2 "github.com/archway-network/archway/app/upgrades/4_0_2"
 )
 
 // UPGRADES
@@ -22,7 +22,7 @@ var Upgrades = []upgrades.Upgrade{
 	upgrade2_0_0.Upgrade,      // v2.0.0
 	upgrade3_0_0.Upgrade,      // v3.0.0
 	upgrade4_0_0.Upgrade,      // v4.0.0
-	upgrade4_0_1.Upgrade,      // v4.0.1
+	upgrade4_0_2.Upgrade,      // v4.0.2
 }
 
 func (app *ArchwayApp) setupUpgrades() {