+++ interchaintest/setup.go
@@ -8,8 +8,8 @@ import (
 )
 
 const (
-	initialVersion = "v4.0.0" // The last release of the chain. The one the mainnet is running on
-	upgradeName    = "v4.0.1" // The next upgrade name. Should match the upgrade handler.
+	initialVersion = "v4.0.1" // The last release of the chain. The one the mainnet is running on
+	upgradeName    = "v4.0.2" // The next upgrade name. Should match the upgrade handler.
 	chainName      = "archway"
 )