+++ CHANGELOG.md
@@ -28,7 +28,11 @@ Contains bug fixes.
 Contains all the PRs that improved the code without changing the behaviours.
 -->
 
-## [main]
+## [v4.0.2]
+
+### Changed
+
+- [#440](https://github.com/archway-network/archway/pull/440) - Retagging with v4.0.2 to prevent dual tagging of same commit and same tag name
 
 ### Fixed
 
@@ -38,7 +42,7 @@ Contains all the PRs that improved the code without changing the behaviours.
 
 ### Fixed
 
-- [#437](https://github.com/archway-network/archway/pull/437) - adding upgrade handler with missing burn permissions for feecollector account
+- [#437](https://github.com/archway-network/archway/pull/437) - Adding upgrade handler with missing burn permissions for feecollector account
 
 ## [v4.0.0]